Edited RegisterViewModel and introduced two new validation attributes
RegisterViewModel now has BirthDate, First and LastName, PhoneNumber, BirthDate. Also introduced new CustomStringLength attribute which does the same as StringLength but it works with custom not constant messages. DateAttribute validates the date. All the error messages are in bulgarian because the app will be translated to bulgarian later in development